+++ src/main.jsx
@@ -2,6 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
 import { initTema } from './lib/tema.js'
+import { pulihkanDariCermin } from './lib/cermin.js'
 import './index.css'
 
 initTema()
@@ -13,8 +14,16 @@ document.addEventListener('contextmenu', (e) => {
   }
 })
 
-ReactDOM.createRoot(document.getElementById('root')).render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-)
+// Sebelum render: kalau localStorage kosong tapi ada cermin data di Documents,
+// pulihkan dulu lalu muat ulang agar data warung tidak hilang saat "clear data".
+pulihkanDariCermin().then((dipulihkan) => {
+  if (dipulihkan) {
+    location.reload()
+    return
+  }
+  ReactDOM.createRoot(document.getElementById('root')).render(
+    <React.StrictMode>
+      <App />
+    </React.StrictMode>,
+  )
+})